corrección en migración cuando ya existe la columna

+++ modules/Payroll/Database/Migrations/2019_04_11_162805_add_field_gender_id_to_payroll_Staffs_table.php
@@ -14,10 +14,12 @@ class AddFieldGenderIdToPayrollStaffsTable extends Migration
     public function up()
     {
         Schema::table('payroll_staffs', function (Blueprint $table) {
-            $table->integer('payroll_gender_id')->unsigned()
-                  ->comment('identificador de género que pertenece al personal');
-            $table->foreign('payroll_gender_id')->references('id')->on('payroll_genders')
-                  ->onDelete('restrict')->onUpdate('cascade');
+            if (!Schema::hasColumn('payroll_staffs', 'payroll_gender_id')) {
+                $table->integer('payroll_gender_id')->unsigned()
+                      ->comment('identificador de género que pertenece al personal');
+                $table->foreign('payroll_gender_id')->references('id')->on('payroll_genders')
+                      ->onDelete('restrict')->onUpdate('cascade');
+            }
         });
     }
